+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add multiple selection variants of the widgets
 - fix widget not rendering initially selected choices
+- reworked demo
 
 ## 0.4.0 (2023-07-10)