+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased]
+## 0.5.0 (2023-07-17)
 
 - add multiple selection variants of the widgets
 - fix widget not rendering initially selected choices